updated db table setup
